Switch client config from YAML to JSON
This corresponds to the change in packaging from providing the old
YAML client config to the new JSON client config, which includes some
minor structural and naming changes to the configuration format. Most
notably, the config is now intended to be end-user-editable, so page
descriptors and bundle descriptors have been removed from the config.

Also adds support for specifying limits in the bundle options,
although they are currently ignored by the implementations.