Convert CamelCase method name to snake_case
The currently, the code generator uses Guava's CaseFormat to convert
the CamelCase method names defined in the IDL to more Pythonic
snake_case. However, this conversion is not reflected in the client
config, since that config is language-independent. Instead, we have
GAX do the conversion in a format that is intended to match the Guava
case formatting exactly.

(Note that it is necessary for Guava CaseFormat and GAX to provide
identical behavior, or else the client config will not be parsed
properly.)