Fix bundling bug (#114)
Previously, one timer was associated with each Executor. This timer
was not reset when it fired, which meant that only a single request
could ever be triggered by a delay threshold.

This change switches to using one timer per bundle. This correctly
allows different bundles to fire under different timers.