Dependency checking environment (#124)
* Dependency checking environment

Add an tox environment that does static analysis (via pylint) without
install test-requirements.txt; this ensures that the dependencies
specified in setup.py are enough. Previously, the
tests could pass but the package might not build if a necessary
dependency was specified in the test environment but not in the real
package metadata. Remove requirements.txt, which is redundant with
setup.py.

Also, switch to using grpcio-tools to provide protoc
- It is difficult to debug linuxbrew problems on Travis; it's cleaner just
  to install a Python package.
- Note that grpcio-tools does not install properly for pip versions <8.

Also, fix an import for style
- Directly importing the method is inconsistent with the rest
  of this repo.
- The import causes Sphinx to document the method as though it were
  part of `api_callable`, even though it really belongs to a third-
  party library.